Allow incomplete player participations
I think the best way to handle deleting of characters and players
already in use in games is to allow games to be incomplete and consider
this when e.g. deriving any stats from them. The frontend will display a
warning when a game is incomplete. But I think thats more convenient
then some form of cascading delete, or blocking a delete until the
character/player is not used anywhere anymore.